Move Configurator::Question into its own file
It is too long to live in the same file as the Configurator. Ideally it
would be refactored into QuestionTree
